Add test for unobserved schemes
The script is attempting to ensure backwards compatibility while but it could easily
break with new changes to pubMLST.

Adding this support will notify the user of any unobserved schemes that were expected.

Changes can then be caught.

Requires a file with a list of expected schemes to be in the same folder
as the script.